+++ lib/main.dart
@@ -1,120 +1,115 @@
 import 'package:flutter/material.dart';
 
 import './questions.dart';
+import 'screens/startscreen.dart';
 import 'layoutUI/quiz.dart';
 import 'layoutUI/results.dart';
-import 'widgets/quizbutton.dart';
-
 
 void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-
-
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      theme: ThemeData(
+        theme: ThemeData(
           primaryColor: Colors.grey[900],
           accentColor: Colors.grey[700],
           canvasColor: Colors.grey[850],
         ),
-      routes: <String, WidgetBuilder>{
-        'Big Five Quiz' : (context) => MyBiGFiveQuizApp(),
-      },
-      home: StartScreen()
-    );
+        routes: <String, WidgetBuilder>{
+          'Big Five Quiz': (context) => MyBiGFiveQuizApp(),
+        },
+        home: StartScreen());
   }
 }
 
-class StartScreen extends StatelessWidget {
+// class StartScreen extends StatelessWidget {
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: Colors.grey[850],
-        body: Container(
-          alignment: Alignment.center,
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-  crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Text('THE',
-              style: TextStyle(
-                
-                fontSize: 100,
-                color: Colors.grey[850],
-                fontWeight: FontWeight.bold,
-                shadows: [
-                  Shadow(
-                    color: Colors.grey[800],
-                    offset: Offset(-5, -5),
-                    blurRadius: 10,
-                    ),
-                  Shadow(
-                    color: Colors.grey[900],
-                    offset: Offset(5, 5),
-                    blurRadius: 10,
-                  )
-                ],
-              ),),
-              Text('BIG',
-              style: TextStyle(
-                fontSize: 100,
-                color: Colors.grey[850],
-                fontWeight: FontWeight.bold,
-                shadows: [
-                  Shadow(
-                    color: Colors.grey[800],
-                    offset: Offset(-5, -5),
-                    blurRadius: 10,
-                    ),
-                  Shadow(
-                    color: Colors.grey[900],
-                    offset: Offset(5, 5),
-                    blurRadius: 10,
-                  )
-                ],
-              ),
-              ),
-              Text('FIVE',
-              style: TextStyle(
-                fontSize: 100,
-                color: Colors.grey[850],
-                fontWeight: FontWeight.bold,
-                shadows: [
-                  Shadow(
-                    color: Colors.grey[800],
-                    offset: Offset(-5, -5),
-                    blurRadius: 10,
-                    ),
-                  Shadow(
-                    color: Colors.grey[900],
-                    offset: Offset(5, 5),
-                    blurRadius: 10,
-                  )
-                ],
-              ),
-              ),
-              Padding(padding: EdgeInsets.only(top:20,)),
-              QuizButton(
-                buttonText: 'START QUIZ',
-                buttonFunction: (){
-                  Navigator.pushNamed(
-      context,
-      'Big Five Quiz'
-  ); //'Big Five Quiz'
-                },)
-            ],
-          ),
-        ),
-      );
-  }
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//         backgroundColor: Colors.grey[850],
+//         body: Container(
+//           alignment: Alignment.center,
+//           child: Column(
+//             mainAxisAlignment: MainAxisAlignment.center,
+//   crossAxisAlignment: CrossAxisAlignment.center,
+//             children: <Widget>[
+//               Text('THE',
+//               style: TextStyle(
 
-}
+//                 fontSize: 100,
+//                 color: Colors.grey[850],
+//                 fontWeight: FontWeight.bold,
+//                 shadows: [
+//                   Shadow(
+//                     color: Colors.grey[800],
+//                     offset: Offset(-5, -5),
+//                     blurRadius: 10,
+//                     ),
+//                   Shadow(
+//                     color: Colors.grey[900],
+//                     offset: Offset(5, 5),
+//                     blurRadius: 10,
+//                   )
+//                 ],
+//               ),),
+//               Text('BIG',
+//               style: TextStyle(
+//                 fontSize: 100,
+//                 color: Colors.grey[850],
+//                 fontWeight: FontWeight.bold,
+//                 shadows: [
+//                   Shadow(
+//                     color: Colors.grey[800],
+//                     offset: Offset(-5, -5),
+//                     blurRadius: 10,
+//                     ),
+//                   Shadow(
+//                     color: Colors.grey[900],
+//                     offset: Offset(5, 5),
+//                     blurRadius: 10,
+//                   )
+//                 ],
+//               ),
+//               ),
+//               Text('FIVE',
+//               style: TextStyle(
+//                 fontSize: 100,
+//                 color: Colors.grey[850],
+//                 fontWeight: FontWeight.bold,
+//                 shadows: [
+//                   Shadow(
+//                     color: Colors.grey[800],
+//                     offset: Offset(-5, -5),
+//                     blurRadius: 10,
+//                     ),
+//                   Shadow(
+//                     color: Colors.grey[900],
+//                     offset: Offset(5, 5),
+//                     blurRadius: 10,
+//                   )
+//                 ],
+//               ),
+//               ),
+//               Padding(padding: EdgeInsets.only(top:20,)),
+//               QuizButton(
+//                 buttonText: 'START QUIZ',
+//                 buttonFunction: (){
+//                   Navigator.pushNamed(
+//       context,
+//       'Big Five Quiz'
+//   ); //'Big Five Quiz'
+//                 },)
+//             ],
+//           ),
+//         ),
+//       );
+//   }
+
+// }
 
 class MyBiGFiveQuizApp extends StatefulWidget {
   @override
@@ -181,36 +176,38 @@ class MyBiGFiveQuizAppState extends State<MyBiGFiveQuizApp> {
     });
   }
 
-void returnToStart() {
-  Navigator.pop(context);
-}
+  void returnToStart() {
+    Navigator.pop(context);
+  }
 
   //Initializing build//
 
   @override
   Widget build(BuildContext context) {
-    return 
-        Scaffold(
-          appBar: AppBar(
-            title: Text('BIG FIVE Personality Quiz'),
-          ),
-          body: questionIndex < questions.length
-              ? SingleChildScrollView(
-                  scrollDirection: Axis.vertical,
-                  child: Column(
-                    children: <Widget>[
-                      Quiz(
-                        questionIndex: questionIndex,
-                        d: nextQuestionD,
-                        sd: nextQuestionSD,
-                        n: nextQuestionN,
-                        sa: nextQuestionSA,
-                        a: nextQuestionA,
-                      ),
-                    ],
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('BIG FIVE Personality Quiz'),
+      ),
+      body: questionIndex < questions.length
+          ? SingleChildScrollView(
+              scrollDirection: Axis.vertical,
+              child: Column(
+                children: <Widget>[
+                  Quiz(
+                    questionIndex: questionIndex,
+                    d: nextQuestionD,
+                    sd: nextQuestionSD,
+                    n: nextQuestionN,
+                    sa: nextQuestionSA,
+                    a: nextQuestionA,
                   ),
-                )
-              : Results(restartQuiz: resetQuiz, goHome: returnToStart, ),
-        );
+                ],
+              ),
+            )
+          : Results(
+              restartQuiz: resetQuiz,
+              goHome: returnToStart,
+            ),
+    );
   }
 }
